messanger/requirements.txt
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.1.0
boto3==1.10.17
botocore==1.13.17
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
defusedxml==0.6.0
Django==3.0.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-storages==1.7.2
djangorestframework==3.11.0
docutils==0.15.2
factory-boy==2.12.0
Faker==4.0.1
flamegraph==0.1
freezegun==0.3.15
idna==2.8
importlib-metadata==0.23
jmespath==0.9.4
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.2
Pillow==6.2.1
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.3
python-dateutil==2.8.0
python-magic==0.4.15
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
selenium==3.141.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
